Shell for show special regs, and include updates:
- Include updates:
Changed to allow including both data segment files
and code segment files.
This imposes a particular style for main programs
that will use includes.
See the shell code for details.

TODO: rework the make files to support the include
system logic.

- Show special registers
A shell of a program that will show (dump to sysout)
the contents of all of the special registers.

TODO: Add logic that builds special register contents
and writes same to sysout.